Upsert based on ID field
