+++ packages/twenty-front/src/modules/workflow/workflow-steps/workflow-actions/components/WorkflowEditActionUpdateRecord.tsx
@@ -198,6 +198,7 @@ export const WorkflowEditActionUpdateRecord = ({
             }
             readonly={isFormDisabled ?? false}
             defaultFields={formData.fieldsToUpdate}
+            actionType="UPDATE_RECORD"
           />
         )}
